backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
openai==1.12.0
httpx==0.26.0
pydantic-settings==2.1.0
pyttsx3==2.90
python-multipart==0.0.9
